requirements.txt
Django~=5.0.4
requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
opengraph~=0.5
djangorestframework~=3.15.1
Faker~=24.11.0